requirements.txt
paddleocr
paddlepaddle
pillow
numpy
requests
mysql-connector-python
fastapi
python-jose
bcrypt
python-dotenv
uvicorn